backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.4",
    "django-extensions>=3.2.3",
    "django-unfold>=0.44.0",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
]

[tool.ruff]
line-length = 120
exclude = ["*/**/migrations/*"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
]
